refactor(Slider): 重构代码结构
